Do not malloc() the first hash nodes unless in 'add' mode
It is not necessary to allocate the head hash leaves when the
program is not adding an image to the pile. Read mode does not use
hashes so this would wastefully allocate a load of memory for no
reason that would never be touched when reading an image from the
image pile.